Add files via upload
Hello! This is Neeti the legal assistant.
She has answers to all your law-related queries. She is able to converse in both textual as well as spoken formats.
She will soon be able to converse in multilingual languages. 
She is integrated into a website that is made in collaboration with my classmates.